first final version v1
v1.0.0

+++ registration_and_login/delete_user.php
@@ -20,25 +20,28 @@ if ($conn->connect_error) {
 if (isset($_POST['delete'])) {
 
     $name = $_POST['name'];
-    
-    $sql = "DELETE FROM user_info WHERE name='$name'";
 
-    if (mysqli_query($conn, $sql)) {
-        echo "<h1>User deleted successfully</h1>";
-      } else {
-        echo "Error deleting record: " . mysqli_error($conn);
-      }
-      
+    $sql2 = "SELECT * FROM user_info WHERE name='$name'";
 
-     }   
+    $result = mysqli_query($conn, $sql2);
+
+    if (mysqli_num_rows($result) > 0) {
+
+        $sql = "DELETE FROM user_info WHERE name='$name'";
+        if (mysqli_query($conn, $sql)) {
+            echo "<h1>Account deleted successfully</h1>";
+        } else {
+            echo "Error deleting record: " . mysqli_error($conn);
+        }
+
+    }else {
+        echo "<h1> User does not exist </h1>";
+        // header("Location: index.php?error=Incorect User name or password");
+    }
+}   
 
         
     
-
-
-
-
-
 
 
 $conn->close();
@@ -51,7 +54,7 @@ $conn->close();
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>حذف طبيب</title>
+    <title>Remove account</title>
 
     <!-- Font Icon -->
     <link rel="stylesheet" href="fonts/material-icon/css/material-design-iconic-font.min.css">
@@ -73,17 +76,17 @@ $conn->close();
                     </div>
 
                     <div class="signin-form">
-                        <h2 class="form-title"> حذف مريض</h2>
+                        <h2 class="form-title">Remove patient account</h2>
                         <form method="POST" class="register-form" id="login-form">
                             <div class="form-group">
                                 <label for="name"><i class="zmdi zmdi-account material-icons-name"></i></label>
-                                <input type="text" name="name" id="name" placeholder="اسم المستخدم" required />
+                                <input type="text" name="name" id="name" placeholder="Patient name" required />
                             </div>
                       
                             
                             <div class="form-group form-button">
                                 <input type="submit" name="delete" id="signin" class="form-submit"
-                                    value="حذف " />
+                                    value="Remove" />
                             </div>
                         </form>
                     </div>
